Move shortcut summary into its own file
